+++ apps/proxy/urls.py
@@ -8,6 +8,5 @@ urlpatterns = [
     path('stats/', stats_views.combined_stats, name='combined_stats'),
     path('ts/', include('apps.proxy.live_proxy.urls')),
     path('catchup/', include('apps.timeshift.urls')),
-    path('hls/', include('apps.proxy.hls_proxy.urls')),
     path('vod/', include('apps.proxy.vod_proxy.urls')),
 ]